Added Documentation for Contributing
I added basic documentation for contributing to the project. This
includes a basic overview of the Git architecture and how we organize
the branches based on the image Martin provided to me. I also included
a basic set of rules and expectations that can be expanded upon by
the current EBoard and Martin.

I provided a section for the expectation of Code, and Documentation
That is how should they write the code (Linting?) and how they should
write comments and documentation.

I Organized this into a separate folder and README to prevent conflict.