Typos and Organization
Organized the Images in the Main readme into a "Images" folder to clean
it up. Additionally I also changed the references to the images in the
code to reflect the local path rather then the URL, this is so the other
branches that make changes to the images, and logo will have them
reflected rather than just pulling the logo on the main branch.

Fixed Typo in the Contribution Guidelines and added a temporary
statement for code guidelines (i.e refer to Martin/EBoard).

Fixed Typo in the Main README, and modified the section containing the
Contributors list to use a better set of words (I think).

"These are a list" -> "This is a list"